% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-stream-info.R
\name{stream-metadata}
\alias{stream-metadata}
\alias{streamName}
\alias{streamType}
\alias{streamChannels}
\alias{streamRate}
\alias{streamDtype}
\title{Stream metadata accessors}
\usage{
streamName(x)

streamType(x)

streamChannels(x)

streamRate(x)

streamDtype(x)
}
\arguments{
\item{x}{A \code{StreamInfo}, endpoint, or ring buffer.}
}
\value{
The requested metadata field.
}
\description{
Stream metadata accessors
}
\examples{
info <- streamInfo("eeg-demo", type = "EEG",
                   channel_names = c("C3", "C4"), nominal_srate = 250,
                   dtype = "float32")
streamName(info)
streamType(info)
streamChannels(info)
streamRate(info)
streamDtype(info)
}
