% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-stream.R
\docType{class}
\name{StreamSink-class}
\alias{StreamSink-class}
\title{Virtual stream sink}
\description{
Transport backends extend this class and the stream-operation generics.
}
\examples{
# StreamSink is virtual; transport outlets such as lslOutlet() extend it.
isVirtualClass("StreamSink")
}
