% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-stream.R
\docType{class}
\name{StreamSource-class}
\alias{StreamSource-class}
\title{Virtual stream source}
\description{
Transport backends extend this class and the stream-operation generics.
}
\examples{
# StreamSource is virtual; loopbackSource() returns a concrete instance.
info <- streamInfo("demo", type = "EEG",
                   channel_names = "C3", nominal_srate = 100)
is(loopbackSource(info), "StreamSource")
}
